feat: Admin page (art pieces table)
Implemented a table for administators to view all of the art in the database.
Admins can now accept and decline art pieces and also have it updated in the database.
There is also the option to view the table based on specific status (ex: only view Accepted art or only view pending art)

Addresses: Issue #6