feat: Update request form, icons on comments
Modified the request form to include a "Confirm Email" field to ensure that
the users are sending the correct email.

If an artist makes a comment on their own art piece, it will now show a special icon
to show that the commenter is the artisit of the art piece.

Addresses: Issues #17, #18, #3