perf: Update authState to include more values
The authState variable now contains the first name and last name of the user.
Initially, a query was performed in the gallery page (for the comments section) and the
"Request Art" page just to retrieve the name of the current user that's logged in and display it.
Now we the first name, last name, username, and account type is stored in a variable with the useContext
React Hook. This is better because the query to retrieve the user's info is only performed once when after
they log in. The auth/info get request route was able to be removed and all other variables used for the values
that's already stored in the authState variable.