fix: Netlify Page Not Found Error
Trying a solution that should fix the "Page Not Found" error for the deployed Netlify app.
This error would occur upon refreshing the page or clicking on other valid links in the website.
I included a _redirects file in the public folder which should give Netlify the support of client-side routing.